Begin refactoring preloading; Fix menu loading
Menu was getting double loaded from the preloading and from
StartManager, so I had to finish reverting the load position changes.
I've also moved preloading into its own method as it was getting to be
more than half of the mod loading. Also swapped the usage of bools to
waiting on all the coroutines to finish via yield return, which
simplified loading.