Fix invalidation of short branches in PrePatcher
When multiple PlayerData field references are within a method, a short
branch can have its operand changed such that the offset it points to is
greater than 1-byte. As a result of this, we need to patch such branches
into their non-short versions to ensure that the method is not
invalidated on module write, which causes the short branch operands to
become null. In addition, this also creates the Output folder for the
PrePatcher, which was a missing step in the build process that had to be
done by hand earlier.