Use ReflectionHelper for get/set cstate
I don't really think it's necessary to have a modhook for this - since the cstate isn't saved, there's not really a reason to modify the result of the methods instead of simply modifying the cstate directly.

Actually that's not totally true - you may want to be able to log changes to the cstate. But I don't think it's worth the hassle of modifying the prepatcher just for this specific use case.