Make AfterTakeDamage a bit more consistent
AfterTakeDamage now is raised with a damage value of 0 (rather than the original) if they were protected by Baldur Shell. The new application, while the player is invulnerable, is only raised when they were going to take damage anyway (because the damage source is spikes or acid) - in normal gameplay, hitting an enemy (hazard type = 1) causes several extra "damage" instances, which get filtered through the invulnerable section of the routine where nothing happens because it's hazard type 1, so it's dishonest of us to raise the event.